Switch to Webpack, add DebugPanel
Switched from Parcel to Webpack and made a starting config for us that should be equivalent.
There were some changes to the `index.html` because Webpack's entry is the script file whereas Parcel's
entry was the html file.

`index.html` is now used as a template and Webpack injects the bundled script and stylesheet.

The DebugPanel just shows the "frame time" right now, which is what I changed the FPS counter to in the Renderer.

The Renderer now tracks how many milliseconds the last frame took to render, so we can use that variable in the debug panel.
This is independent of the `requestAnimationFrame` loop, so the delay between frame requests is not included in this value.

I also set all text in the body of the page to be monospace.