Add resource / asset loading from JSON config files
This is a big change that affects lots of configuration.

The biggest API change is that you can now load resources via their `.load(url)`
static methods.

Creating new Materials is much more involved than it was previously.
Take this diff as an example of how to do that.

I added path-browserify to deal with path resolution in the JSON configs

I set the tsconfig target to esnext again and enabled top level async so
we can async load Shaders in Material subclasses.

Removed some rules for assets in the webpack.config.js and added
the CopyPlugin to move all assets to the dist folder. This
will make sure our asset URLs are persisted from src after build.

I used the new Model load method in the script.js:30 as an example.